+++ crates/ruma-common/src/events/enums.rs
@@ -40,11 +40,8 @@ event_enum! {
         "m.call.invite" => super::call::invite,
         "m.call.hangup" => super::call::hangup,
         "m.call.candidates" => super::call::candidates,
-        #[cfg(feature = "unstable-msc2746")]
         "m.call.negotiate" => super::call::negotiate,
-        #[cfg(feature = "unstable-msc2746")]
         "m.call.reject" => super::call::reject,
-        #[cfg(feature = "unstable-msc2746")]
         "m.call.select_answer" => super::call::select_answer,
         #[cfg(feature = "unstable-msc3954")]
         #[ruma_enum(alias = "m.emote")]
@@ -340,9 +337,10 @@ impl AnyMessageLikeEventContent {
             }
             #[cfg(feature = "unstable-msc3381")]
             Self::PollStart(_) => None,
-            #[cfg(feature = "unstable-msc2746")]
-            Self::CallNegotiate(_) | Self::CallReject(_) | Self::CallSelectAnswer(_) => None,
-            Self::CallAnswer(_)
+            Self::CallNegotiate(_)
+            | Self::CallReject(_)
+            | Self::CallSelectAnswer(_)
+            | Self::CallAnswer(_)
             | Self::CallInvite(_)
             | Self::CallHangup(_)
             | Self::CallCandidates(_)